+++ src/shared/lib/botDetector.ts
